+++ logic/Gaming/Game.cs
@@ -150,7 +150,7 @@ namespace Gaming
 
                                 finally
                                 {
-                                    gameMap.GameObjLockDict[GameObjType.Character].ExitReadLock();
+                                    gameMap.GameObjLockDict[GameObjType.Generator].ExitReadLock();
                                 }
                             },
                         timeInterval: GameData.checkInterval,
@@ -179,11 +179,17 @@ namespace Gaming
 
             propManager.StartProducing();
 
-            // 开始游戏
-            if (!gameMap.Timer.StartGame(milliSeconds))
-                return false;
+            new Thread
+            (
+                () =>
+                {
+                    if (!gameMap.Timer.StartGame(milliSeconds))
+                        return;
 
-            EndGame();  // 游戏结束时要做的事
+                    EndGame();  // 游戏结束时要做的事
+                }
+            )
+            { IsBackground = true }.Start();
 
             return true;
         }
@@ -333,7 +339,7 @@ namespace Gaming
             Character? player = gameMap.FindPlayer(playerID);
             if (player != null)
             {
-                return skillManager.UseActiveSkill(this.GameMap, player, activeSkillType);
+                return skillManager.UseActiveSkill(player, activeSkillType);
             }
             else
                 return false;
@@ -348,7 +354,7 @@ namespace Gaming
             {
                 foreach (Character player in gameMap.GameObjDict[GameObjType.Character])
                 {
-                    skillManager.UseAllPassiveSkill(this.GameMap, player);
+                    skillManager.UseAllPassiveSkill(player);
                 }
             }
             finally
@@ -439,10 +445,10 @@ namespace Gaming
                 teamList.Add(new Team());
             }
 
-            skillManager = new SkillManager();
             attackManager = new AttackManager(gameMap);
             actionManager = new ActionManager(gameMap);
             propManager = new PropManager(gameMap);
+            skillManager = new SkillManager(gameMap, actionManager, attackManager, propManager);
         }
     }
 }
