+++ logic/规则Logic.md
@@ -5,6 +5,7 @@
 - 该规则直接服务于Sever,并非选手版本
 - *斜体表示Logic底层尚未（完全）实现*
 - []表示待决定
+- ~~表示暂不实现~~
 
 ## 游戏简介
 - 1位监管者对抗4位求生者的非对称竞技模式
@@ -90,11 +91,11 @@
         IsStunned = 11,
         IsTryingToAttack = 12,//指前摇
         IsLockingTheDoor = 13,
-        IsRummagingInTheDrawer = 14,
+        IsRummagingInTheChest = 14,
         IsClimbingThroughWindows = 15,
     }
   ~~~
-- *Bgm(字典)*
+- Bgm(字典)
 - 得分
 - ~~回血率/原始回血率~~
 - 当前子弹类型
@@ -108,15 +109,15 @@
 - 拥有的被动技能(列表)
 - 拥有的主动技能(列表)
 - 各个主动技能CD(字典)
-- *警戒半径*
-- *double 隐蔽度*
-- *翻墙速度*
-- *开锁门速度*
+- 警戒半径
+- double 隐蔽度
+- *翻窗时间*
+- 开锁门时间
 
 ### 学生：人物
 - 修理电机速度
 - 治疗速度
-- 救人速度
+- ~~救人速度~~
 - ~~自愈次数~~
 - 沉迷游戏程度
 - 最大沉迷游戏程度
@@ -171,6 +172,7 @@
 
 ### 箱子：物体
 - *里面的道具*
+- *是否开启*
 
 ### 门：物体
 - *属于那个教学区*
@@ -225,7 +227,7 @@
 - *翻越窗户是一种交互行为，执行时，实质是限定方向的减速运动*
 
 ### 箱子
-- *监管者和求生者都能与箱子交互。*
+- *监管者和求生者都能与箱子交互,同一时刻就允许一人进行翻找*
 - *开启箱子有不同概率获得不同道具。*
 - *搜寻物品的基础持续时间为10秒。*
 - *未搜寻完成的箱子在下一次需要重新开始搜寻。*
